Move sink output to SPARQL
